Extract `NumericDataTest` to `test/cases/numeric_data_test.rb`
To ease to find the numeric data tests, extract `NumericDataTest` to
`test/cases/numeric_data_test.rb` dedicated file.
